+++ src/main/java/com/hepi/music_api/security/user/repository/UserRepository.java
@@ -13,12 +13,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 @Repository
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
   Optional<User> findByEmail(String email);
 
   Page<User> findAll(Pageable pageable);
-  Optional<User> findById(Integer  usercode);
+  Optional<User> findById(Long  usercode);
 
   @Transactional
   @Modifying
